:art:Extract DTO class
