+++ backend-fastapi/dto.py
@@ -1,0 +1,10 @@
+"""DTO modules."""
+from pydantic import BaseModel
+from datetime import datetime
+
+
+class WeatherForecast(BaseModel):
+    date: datetime
+    temperature_c: int
+    temperature_f: int
+    summary: str
